order service sends log events to kafka

+++ services/order-service/order-service/Order/OrderService.cs
@@ -1,6 +1,8 @@
+using System.Text.Json;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using order_service.Inventory;
+using order_service.Logging;
 using Steeltoe.Discovery;
 using Steeltoe.Discovery.Eureka;
 
@@ -27,7 +29,7 @@ public class OrderService
         foreach (var productId in orderInput.ProductIds)
         {
             var discoveryProduct =
-                await Inventory.InventoryEurekaOps.GetProductFromInventory(productId, discoveryClient, client);
+                await InventoryEurekaOps.GetProductFromInventory(productId, discoveryClient, client);
             if (discoveryProduct == null)
                 return new NotFound<Order>($"The following product id does not exist: {productId}");
             if (discoveryProduct.Product.Stock <= 0)
@@ -35,10 +37,15 @@ public class OrderService
             Thread.Sleep(250);
             InventoryEurekaOps.DecreaseStockByOne(productId, discoveryClient, client);
             Thread.Sleep(250);
-
         }
 
         var order = await orderRepository.CreateOrder(orderInput);
+
+        var message = JsonSerializer.Serialize(order);
+        await LoggingEurekaConnection.LogMessage(new Logging.Logging("INFO",
+                $"This order has been created: {message}", DateTime.Now, order.Email),
+            discoveryClient, client);
+
         return new Ok<Order>(order);
     }
 
@@ -75,16 +82,30 @@ public class OrderService
         if (id == Guid.Empty)
             return new BadRequest<bool>("id cannot be empty");
 
-        var orders = await orderRepository.GetOrderById(id);
-        orders.ProductIds.ForEach(async x =>
+        var order = await orderRepository.GetOrderById(id);
+        if (order == null)
+        {
+            await LoggingEurekaConnection.LogMessage(new Logging.Logging("CRITICAL",
+                $"Order with id {id} cannot be deleted", DateTime.Now, "unknown"), discoveryClient, client);
+            return new Conflict<bool>("Order could not be deleted");
+        }
+
+        order.ProductIds.ForEach(async x =>
         {
             Thread.Sleep(250);
             await InventoryEurekaOps.IncreaseStockByOne(x, discoveryClient, client);
         });
         var deleted = await orderRepository.DeleteOrder(id);
 
-        return deleted
-            ? new NoContent<bool>($"Deleted entry with id: {id}")
-            : new Conflict<bool>("Order could not be deleted");
+        if (!deleted)
+        {
+            await LoggingEurekaConnection.LogMessage(new Logging.Logging("CRITICAL",
+                $"Order with id {id} cannot be deleted", DateTime.Now, order.Email), discoveryClient, client);
+            return new Conflict<bool>("Order could not be deleted");
+        }
+
+        await LoggingEurekaConnection.LogMessage(new Logging.Logging("INFO",
+            $"Order with id: {id} was deleted", DateTime.Now, order.Email), discoveryClient, client);
+        return new NoContent<bool>($"Deleted entry with id: {id}");
     }
 }
